Fix bug in MQTTS topic registration (issue #5)
If a MQTTS client missed a SUBACK and consequently resent the SUBSCRIBE,
rsmb would send the next SUBACK with a different topic id. But later
PUBLISHes on that topic would use the first topic id which was unknown
to the client.

Fix is to use the same check for a previously registered topic in
MQTTSProtocol_handleSubscribes that MQTTSProtocol_handleRegisters uses.